Adds a name parameter to the add_arg() Job method.
This allows for user-defined log, error, and output files for each
arg in a Job. If 'name' is not specified for an arg, then it will
go to the main log, error, and output files for that Job. However,
if 'name' is specified, then a separate log, error, and output file
will be generated specifically for that node.